+++ src/utils/validation/email.ts
@@ -1,4 +1,4 @@
-import { errorMsg } from '../variables';
+import { errorMsg } from '../../constant';
 
 export function emailValidation(value: string): string {
   let error: string = '';
