+++ src/utils/validation/password.ts
@@ -1,4 +1,4 @@
-import { errorMsg } from '../variables';
+import { errorMsg } from '../../constant';
 
 export function passwordValidation(value: string): string {
   let error: string = '';
